More modifications to ensure this works
I didn't know a means to hit DownloadItem in triggerContentDownload, so previous attempts couldn't be tested properly. I could test it now and I have double-checked that it works by throttling my steam download to 1KBs then subscribing to a mod so it can see that a mod is still in progress of downloading.